adição de restrição por admin

+++ app/controllers/api/v1/game_controller.rb
@@ -1,45 +1,64 @@
 class Api::V1::GameController < ApplicationController
+  acts_as_token_authentication_handler_for User, only: [:create, :update, :destroy]
+  #before_action :authenticate_user!
+
+
+
   def index
     games  = Game.all
     render json: games
   end
-  def create
-    game = Game.new(game_params)
-    if game.save
-      render json: game
-    else
-      render json: {status:'ERROR', message: 'falha ao criar jogo' }, status: :unprocessable_entity
-    end
-  end
 
-  def destroy
-    if Game.exists?(params[:id])
-      game = Game.find(params[:id])
-      if game.destroy!
-        render json: {status:'SUCCESS', message:'jogo foi apagado'}, status: :ok
+
+    def create
+        if current_user.is_Admin
+        game = Game.new(game_params)
+        if game.save
+          render json: game
+        else
+          render json: {status:'ERROR', message: 'falha ao criar jogo' }, status: :unprocessable_entity
+        end
       else
-        render json: {status:'ERROR', message:'falha ao apagar jogo'}, status: :bad_request
-      end
-    else
-      render json: {status: 'ERROR', message: 'falha ao apagar jogo, id não existente'}, status: :bad_request
+        render json: {status:"ERROR", message: 'usuario noa é admin'}, status: :bad_request
 
+      end
     end
 
-  end
-
-  def update
-    if Game.exists?(params[:id])
-      game = Game.find(params[:id])
-      if game.update(game_params)
-        render json: game, status: :ok
+    def destroy
+      if current_user.is_Admin
+        if Game.exists?(params[:id])
+          game = Game.find(params[:id])
+          if game.destroy!
+            render json: {status:'SUCCESS', message:'jogo foi apagado'}, status: :ok
+          else
+            render json: {status:'ERROR', message:'falha ao apagar jogo'}, status: :bad_request
+          end
+        else
+          render json: {status: 'ERROR', message: 'falha ao apagar jogo, id não existente'}, status: :bad_request
+        end
       else
-        render json: {status: 'ERROR', message: 'falha ao editar jogo', data: game}, status: :bad_request
+        render json: {status:"ERROR", message: 'usuario noa é admin'}, status: :bad_request
       end
-    else
-      render json: {status: 'ERROR', message: 'falha ao editar jogo, id não existente'}, status: :bad_request
     end
 
-  end
+    def update
+      if current_user.is_Admin
+        if Game.exists?(params[:id])
+          game = Game.find(params[:id])
+          if game.update(game_params)
+            render json: game, status: :ok
+          else
+            render json: {status: 'ERROR', message: 'falha ao editar jogo', data: game}, status: :bad_request
+          end
+        else
+          render json: {status: 'ERROR', message: 'falha ao editar jogo, id não existente'}, status: :bad_request
+        end
+      else
+      render json: {status:"ERROR", message: 'usuario noa é admin'}, status: :bad_request
+      end
+    end
+
+
   private
   def game_params
     params.require(:game).permit(:name, :price, :category, :publisher, :description, :picture, :trailer)
